+++ defi-alerts/src/app/alerts/create/compound-market/CreateCompoundMarketPage.tsx
@@ -1,38 +1,23 @@
 'use client';
 
 import { AlertCreationResponse, CreateCompoundAlertPayload } from '@/app/api/alerts/create/compound-market/route';
-import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
-import {
-  type Channel,
-  type Condition,
-  type ConditionType,
-  frequencyOptions,
-  type NotificationFrequency,
-  type SeverityLevel,
-  severityOptions,
-} from '@/types/alerts';
+import { type Channel, type Condition, type ConditionType, type NotificationFrequency, type SeverityLevel } from '@/types/alerts';
 import { DoDAOSession } from '@dodao/web-core/types/auth/Session';
 import { useNotificationContext } from '@dodao/web-core/ui/contexts/NotificationContext';
 import { usePostData } from '@dodao/web-core/ui/hooks/fetch/usePostData';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
-import {
-  AlertActionType,
-  ConditionType as PrismaConditionType,
-  DeliveryChannelType,
-  NotificationFrequency as PrismaNotificationFrequency,
-  SeverityLevel as PrismaSeverityLevel,
-} from '@prisma/client';
-import { AlertCircle, ArrowLeft, Bell, ChevronRight, Home, Info, Plus, TrendingUp, X } from 'lucide-react';
-import Link from 'next/link';
+import { AlertActionType, ConditionType as PrismaConditionType, DeliveryChannelType, NotificationFrequency as PrismaNotificationFrequency, SeverityLevel as PrismaSeverityLevel } from '@prisma/client';
+import { ArrowLeft } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
+import { 
+  AlertBreadcrumb, 
+  AlertTypeCard, 
+  MarketSelectionCard, 
+  ConditionSettingsCard, 
+  DeliveryChannelsCard 
+} from '@/components/alerts';
 
 export interface CreateCompoundMarketPageProps {
   session: DoDAOSession;
@@ -43,7 +28,6 @@ export default function CreateCompoundMarketPage({ session }: CreateCompoundMark
   const baseUrl = getBaseUrl();
   const { showNotification } = useNotificationContext();
 
-  console.log('user', session);
   const { postData, loading: isSubmitting } = usePostData<AlertCreationResponse, CreateCompoundAlertPayload>({
     successMessage: 'Your market alert was saved successfully.',
     errorMessage: "Couldn't create alert",
@@ -162,25 +146,15 @@ export default function CreateCompoundMarketPage({ session }: CreateCompoundMark
   const updateCondition = (i: number, field: keyof Condition, val: string) =>
     setConditions((cs) => cs.map((c, idx) => (idx === i ? { ...c, [field]: val } : c)));
 
+  const addCondition = () => setConditions((cs) => [...cs, { conditionType: 'APR_RISE_ABOVE', severity: 'NONE' }]);
+
+  const removeCondition = (i: number) => setConditions((cs) => cs.filter((_, idx) => idx !== i));
+
   const addChannel = () => setChannels((ch) => [...ch, { channelType: 'EMAIL', email: session?.username || '' }]);
 
   const updateChannel = (i: number, field: keyof Channel, val: string) => setChannels((ch) => ch.map((c, idx) => (idx === i ? { ...c, [field]: val } : c)));
 
   const removeChannel = (i: number) => setChannels((ch) => ch.filter((_, idx) => idx !== i));
-
-  // Get contextual message for condition type
-  const getConditionMessage = (conditionType: ConditionType) => {
-    switch (conditionType) {
-      case 'APR_RISE_ABOVE':
-        return 'Alert when APR exceeds your threshold (e.g., alert when APR goes above 5%)';
-      case 'APR_FALLS_BELOW':
-        return 'Alert when APR drops under your threshold (e.g., alert when APR goes below 2%)';
-      case 'APR_OUTSIDE_RANGE':
-        return 'Alert when APR moves outside your specified range (e.g., alert when APR is below 3% or above 6%)';
-      default:
-        return 'Select a condition type to see its description';
-    }
-  };
 
   // Validate form before submission
   const validateForm = () => {
@@ -310,434 +284,50 @@ export default function CreateCompoundMarketPage({ session }: CreateCompoundMark
 
   return (
     <div className="container max-w-6xl mx-auto px-2 py-8">
-      <nav className="flex items-center text-sm mb-6">
-        <Link href="/" className="text-theme-muted hover-text-primary flex items-center gap-1">
-          <Home size={14} />
-          <span>Home</span>
-        </Link>
-        <ChevronRight size={14} className="mx-2 text-theme-muted" />
-        <Link href="/alerts" className="text-theme-muted hover-text-primary flex items-center gap-1">
-          <Bell size={14} />
-          <span>Alerts</span>
-        </Link>
-        <ChevronRight size={14} className="mx-2 text-theme-muted" />
-        <Link href="/alerts/create" className="text-theme-muted hover-text-primary flex items-center gap-1">
-          <TrendingUp size={14} />
-          <span>Create Alert</span>
-        </Link>
-        <ChevronRight size={14} className="mx-2 text-theme-muted" />
-        <span className="text-primary-color font-medium">Compound Market</span>
-      </nav>
+      {/* Breadcrumb */}
+      <AlertBreadcrumb currentPage="Compound Market" />
 
       <div className="mb-8">
         <h1 className="text-3xl font-bold mb-1 text-theme-primary">Create Market Alert</h1>
         <p className="text-theme-muted">Configure a new market alert with your preferred settings.</p>
       </div>
 
-      <Card className="mb-6 border-theme-primary bg-block border-primary-color">
-        <CardHeader className="pb-1">
-          <CardTitle className="text-lg text-theme-primary">Alert Type</CardTitle>
-        </CardHeader>
-
-        <CardContent>
-          <p className="text-sm text-theme-muted mb-4">Choose the type of alert you want to create.</p>
-
-          <RadioGroup value={alertType} onValueChange={(v) => setAlertType(v as 'borrow' | 'supply')} className="space-y-2">
-            {['borrow', 'supply'].map((opt) => (
-              <div key={opt} className="flex items-center space-x-2">
-                <RadioGroupItem value={opt} id={opt} className="h-4 w-4 border border-default rounded-full radio-checked" />
-                <Label htmlFor={opt} className="text-theme-primary label-checked">
-                  {opt === 'borrow' ? 'Borrow Alert' : 'Supply Alert'}
-                </Label>
-              </div>
-            ))}
-          </RadioGroup>
-        </CardContent>
-      </Card>
+      {/* Alert Type */}
+      <AlertTypeCard 
+        alertType={alertType} 
+        setAlertType={setAlertType} 
+      />
 
       {/* Market Selection */}
-      <Card className="mb-6 border-theme-primary bg-block border-primary-color">
-        <CardHeader className="pb-1">
-          <CardTitle className="text-lg text-theme-primary">Market Selection</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <p className="text-sm text-theme-muted mb-4">Select the chains and markets you want to monitor.</p>
-
-          <div className="mb-6">
-            <h3 className="text-md font-medium mb-1 text-theme-primary">Chains</h3>
-            <p className="text-sm text-theme-muted mb-3">Select one or more chains to monitor.</p>
-
-            {/* Chains */}
-            <div className="flex flex-wrap gap-3">
-              {['Ethereum', 'Optimism', 'Arbitrum', 'Polygon', 'Base', 'Unichain', 'Ronin', 'Mantle', 'Scroll'].map((chain) => {
-                const isSel = selectedChains.includes(chain);
-
-                return (
-                  <div
-                    key={chain}
-                    onClick={() => toggleChain(chain)}
-                    className={`rounded-md px-3 py-2 flex items-center cursor-pointer transition-colors border ${
-                      isSel ? 'chip-selected' : 'border-theme-primary'
-                    } ${errors.chains ? 'border-red-500' : ''}`}
-                  >
-                    <div className="chip-checkbox w-4 h-4 rounded border mr-2 flex items-center justify-center">
-                      {isSel && (
-                        <svg width="10" height="8" viewBox="0 0 10 8" fill="none" xmlns="http://www.w3.org/2000/svg">
-                          <path d="M9 1L3.5 6.5L1 4" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
-                        </svg>
-                      )}
-                    </div>
-
-                    <span className="text-theme-primary chip-label">{chain}</span>
-                  </div>
-                );
-              })}
-            </div>
-            {errors.chains && (
-              <div className="mt-2 flex items-center text-red-500 text-sm">
-                <AlertCircle size={16} className="mr-1" />
-                <span>{errors.chains}</span>
-              </div>
-            )}
-          </div>
-
-          {/* Markets */}
-          <div>
-            <div>
-              <h3 className="text-md font-medium mb-1 text-theme-primary">Markets</h3>
-              <p className="text-sm text-theme-muted mb-3">Select one or more markets to monitor.</p>
-
-              <div className="flex flex-wrap gap-3">
-                {['USDC', 'USDS', 'USDT', 'ETH', 'wstETH', 'USDe', 'USDC.e', 'USDbC', 'AERO'].map((market) => {
-                  const isSel = selectedMarkets.includes(market);
-
-                  return (
-                    <div
-                      key={market}
-                      onClick={() => toggleMarket(market)}
-                      className={`rounded-md px-3 py-2 flex items-center cursor-pointer transition-colors border ${
-                        isSel ? 'chip-selected' : 'border-theme-primary'
-                      } ${errors.markets ? 'border-red-500' : ''}`}
-                    >
-                      <div className="chip-checkbox w-4 h-4 rounded border mr-2 flex items-center justify-center">
-                        {isSel && (
-                          <svg width="10" height="8" viewBox="0 0 10 8" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path d="M9 1L3.5 6.5L1 4" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
-                          </svg>
-                        )}
-                      </div>
-
-                      <span className="text-theme-primary chip-label">{market}</span>
-                    </div>
-                  );
-                })}
-              </div>
-              {errors.markets && (
-                <div className="mt-2 flex items-center text-red-500 text-sm">
-                  <AlertCircle size={16} className="mr-1" />
-                  <span>{errors.markets}</span>
-                </div>
-              )}
-            </div>
-          </div>
-        </CardContent>
-      </Card>
+      <MarketSelectionCard
+        selectedChains={selectedChains}
+        toggleChain={toggleChain}
+        selectedMarkets={selectedMarkets}
+        toggleMarket={toggleMarket}
+        errors={errors}
+        showPlatforms={false}
+      />
 
       {/* Condition Settings */}
-      <Card className="mb-6 border-theme-primary bg-block border-primary-color">
-        <CardHeader className="pb-1 flex flex-row items-center justify-between">
-          <CardTitle className="text-lg text-theme-primary">Condition Settings</CardTitle>
-          <Button
-            size="sm"
-            onClick={() => setConditions((cs) => [...cs, { conditionType: 'APR_RISE_ABOVE', severity: 'NONE' }])}
-            className="text-theme-primary border border-theme-primary hover-border-primary hover-text-primary"
-          >
-            <Plus size={16} className="mr-1" /> Add Condition
-          </Button>
-        </CardHeader>
-        <CardContent>
-          <p className="text-sm text-theme-muted mb-4">
-            Define when you want to be alerted about market changes. You will receive an alert if <strong>any</strong> of the following conditions are met for
-            any of the selected markets.
-          </p>
-
-          {/* Conditions List */}
-          {conditions.map((cond, i) => (
-            <div key={i} className="mb-6">
-              <div className="border-t border-primary-color pt-4">
-                {/* Contextual Message */}
-                <div className="mb-4 p-3 bg-theme-secondary rounded-lg border border-theme-primary">
-                  <p className="text-sm text-theme-muted">
-                    <span className="text-primary-color font-medium">Condition {i + 1}:</span> {getConditionMessage(cond.conditionType)}
-                  </p>
-                </div>
-
-                <div className="grid grid-cols-12 gap-4 items-center">
-                  <div className="col-span-1 flex items-center text-theme-muted">
-                    <Badge variant="outline" className="h-6 w-6 flex items-center justify-center p-0 rounded-full text-primary-color">
-                      {i + 1}
-                    </Badge>
-                  </div>
-
-                  {/* Type */}
-                  <div className="col-span-3">
-                    <Select
-                      value={cond.conditionType}
-                      onValueChange={(value) =>
-                        setConditions((cs) =>
-                          cs.map((c, idx) =>
-                            idx === i
-                              ? {
-                                  ...c,
-                                  conditionType: value as ConditionType,
-                                }
-                              : c
-                          )
-                        )
-                      }
-                    >
-                      <SelectTrigger className="w-full hover-border-primary">
-                        <SelectValue placeholder="Select condition type" />
-                      </SelectTrigger>
-                      <SelectContent className="bg-block">
-                        <div className="hover-border-primary hover-text-primary">
-                          <SelectItem value="APR_RISE_ABOVE" className="hover:text-primary-color">
-                            APR rises above threshold
-                          </SelectItem>
-                        </div>
-
-                        <div className="hover-border-primary hover-text-primary">
-                          <SelectItem value="APR_FALLS_BELOW">APR falls below threshold</SelectItem>
-                        </div>
-
-                        <div className="hover-border-primary hover-text-primary">
-                          <SelectItem value="APR_OUTSIDE_RANGE">APR is outside a range</SelectItem>
-                        </div>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  {/* Thresholds */}
-                  {cond.conditionType === 'APR_OUTSIDE_RANGE' ? (
-                    <div className="col-span-4 flex flex-col">
-                      <div className="flex items-center space-x-2">
-                        <Input
-                          type="text"
-                          placeholder="Min (e.g., 3)"
-                          value={cond.thresholdLow || ''}
-                          onChange={(e) => updateCondition(i, 'thresholdLow', e.target.value)}
-                          className={`border-theme-primary focus-border-primary focus:outline-none transition-colors ${
-                            errors.conditions && errors.conditions[i] ? 'border-red-500' : ''
-                          }`}
-                        />
-                        <Input
-                          type="text"
-                          placeholder="Max (e.g., 6)"
-                          value={cond.thresholdHigh || ''}
-                          onChange={(e) => updateCondition(i, 'thresholdHigh', e.target.value)}
-                          className={`border-theme-primary focus-border-primary focus:outline-none transition-colors ${
-                            errors.conditions && errors.conditions[i] ? 'border-red-500' : ''
-                          }`}
-                        />
-                        <span className="text-theme-muted whitespace-nowrap flex-shrink-0">APR</span>
-                      </div>
-                      {errors.conditions && errors.conditions[i] && (
-                        <div className="mt-1 flex items-center text-red-500 text-sm">
-                          <AlertCircle size={14} className="mr-1" />
-                          <span>{errors.conditions[i]}</span>
-                        </div>
-                      )}
-                    </div>
-                  ) : (
-                    <div className="col-span-4 flex flex-col">
-                      <div className="flex items-center">
-                        <Input
-                          type="text"
-                          placeholder={
-                            cond.conditionType === 'APR_RISE_ABOVE'
-                              ? 'Threshold (e.g., 5.0)'
-                              : cond.conditionType === 'APR_FALLS_BELOW'
-                              ? 'Threshold (e.g., 2.0)'
-                              : 'Threshold value'
-                          }
-                          value={cond.thresholdValue || ''}
-                          onChange={(e) => updateCondition(i, 'thresholdValue', e.target.value)}
-                          className={`border-theme-primary focus-border-primary focus:outline-none transition-colors ${
-                            errors.conditions && errors.conditions[i] ? 'border-red-500' : ''
-                          }`}
-                        />
-                        <span className="ml-2 text-theme-muted whitespace-nowrap flex-shrink-0">APR</span>
-                      </div>
-                      {errors.conditions && errors.conditions[i] && (
-                        <div className="mt-1 flex items-center text-red-500 text-sm">
-                          <AlertCircle size={14} className="mr-1" />
-                          <span>{errors.conditions[i]}</span>
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  {/* Severity */}
-                  <div className="col-span-3 flex items-center">
-                    <Select
-                      value={cond.severity === 'NONE' ? undefined : cond.severity}
-                      onValueChange={(value) => setConditions((cs) => cs.map((c, idx) => (idx === i ? { ...c, severity: value as SeverityLevel } : c)))}
-                    >
-                      <SelectTrigger className="w-full hover-border-primary">
-                        <SelectValue placeholder="Severity Level" />
-                      </SelectTrigger>
-                      <SelectContent className="bg-block">
-                        {severityOptions.map((opt) => (
-                          <div key={opt.value} className="hover-border-primary hover-text-primary">
-                            <SelectItem value={opt.value}>{opt.label}</SelectItem>
-                          </div>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                    <TooltipProvider>
-                      <Tooltip>
-                        <TooltipTrigger asChild>
-                          <Button size="icon" className="h-8 w-8 p-0 ml-1 hover-text-primary">
-                            <Info size={16} />
-                          </Button>
-                        </TooltipTrigger>
-                        <TooltipContent className="max-w-xs bg-block p-3 border border-theme-primary">
-                          <p className="text-sm">
-                            Severity level is used for visual indication only. It helps you categorize alerts by importance but doesn’t affect notification
-                            delivery or priority.
-                          </p>
-                        </TooltipContent>
-                      </Tooltip>
-                    </TooltipProvider>
-                  </div>
-
-                  {/* Remove */}
-                  {conditions.length > 1 && (
-                    <Button
-                      variant="ghost"
-                      size="icon"
-                      onClick={() => setConditions((cs) => cs.filter((_, idx) => idx !== i))}
-                      className="col-span-1 text-red-500 h-8 w-8"
-                    >
-                      <X size={16} />
-                    </Button>
-                  )}
-                </div>
-              </div>
-            </div>
-          ))}
-
-          {/* Notification Frequency */}
-          <div className="mt-6">
-            <Label htmlFor="frequency" className="block text-sm font-medium mb-2 text-theme-primary">
-              Notification Frequency
-            </Label>
-            <Select value={notificationFrequency} onValueChange={(value) => setNotificationFrequency(value as NotificationFrequency)}>
-              <SelectTrigger className="w-full hover-border-primary" id="frequency">
-                <SelectValue placeholder="Select frequency" />
-              </SelectTrigger>
-              <SelectContent className="bg-block">
-                {frequencyOptions.map((opt) => (
-                  <div key={opt.value} className="hover-border-primary hover-text-primary">
-                    <SelectItem key={opt.value} value={opt.value}>
-                      {opt.label}
-                    </SelectItem>
-                  </div>
-                ))}
-              </SelectContent>
-            </Select>
-            <p className="text-sm text-theme-muted mt-4">
-              This limits how often you’ll receive notifications for this alert, regardless of how many thresholds are triggered.
-            </p>
-          </div>
-        </CardContent>
-      </Card>
+      <ConditionSettingsCard
+        conditions={conditions}
+        addCondition={addCondition}
+        updateCondition={updateCondition}
+        removeCondition={removeCondition}
+        notificationFrequency={notificationFrequency}
+        setNotificationFrequency={setNotificationFrequency}
+        errors={errors}
+      />
 
       {/* Delivery Channel Settings */}
-      <Card className="mb-6 border-theme-primary bg-block border-primary-color">
-        <CardHeader className="pb-1 flex flex-row items-center justify-between">
-          <CardTitle className="text-lg text-theme-primary">Delivery Channel Settings</CardTitle>
-          <Button size="sm" onClick={addChannel} className="text-theme-primary border border-theme-primary hover-border-primary hover-text-primary">
-            <Plus size={16} className="mr-1" /> Add Channel
-          </Button>
-        </CardHeader>
-        <CardContent>
-          <p className="text-sm text-theme-muted mb-4">Choose how you want to receive your alerts.</p>
-
-          {channels.map((ch, i) => (
-            <div key={i} className="mb-4 flex items-center gap-4">
-              <Select
-                value={ch.channelType}
-                onValueChange={(value) => {
-                  updateChannel(i, 'channelType', value as Channel['channelType']);
-                  if (value === 'EMAIL' && session?.username) {
-                    updateChannel(i, 'email', session.username);
-                  }
-                }}
-              >
-                <SelectTrigger className="w-[150px] hover-border-primary">
-                  <SelectValue placeholder="Select channel" />
-                </SelectTrigger>
-                <SelectContent className="bg-block">
-                  <div className="hover-border-primary hover-text-primary">
-                    <SelectItem value="EMAIL">Email</SelectItem>
-                  </div>
-                  <div className="hover-border-primary hover-text-primary">
-                    <SelectItem value="WEBHOOK">Webhook</SelectItem>
-                  </div>
-                </SelectContent>
-              </Select>
-
-              <div className="flex-1 flex flex-col">
-                {ch.channelType === 'EMAIL' ? (
-                  <>
-                    <Input
-                      type="email"
-                      placeholder="you@example.com"
-                      value={ch.email || ''}
-                      onChange={(e) => updateChannel(i, 'email', e.target.value)}
-                      readOnly={true}
-                      className={`flex-1 border-theme-primary focus-border-primary focus:outline-none transition-colors ${
-                        errors.channels && errors.channels[i] ? 'border-red-500' : ''
-                      } ${ch.channelType === 'EMAIL' ? 'bg-block' : ''}`}
-                    />
-                    {errors.channels && errors.channels[i] && (
-                      <div className="mt-1 flex items-center text-red-500 text-sm">
-                        <AlertCircle size={14} className="mr-1" />
-                        <span>{errors.channels[i]}</span>
-                      </div>
-                    )}
-                  </>
-                ) : (
-                  <>
-                    <Input
-                      type="url"
-                      placeholder="https://webhook.site/..."
-                      value={ch.webhookUrl || ''}
-                      onChange={(e) => updateChannel(i, 'webhookUrl', e.target.value)}
-                      className={`flex-1 border-theme-primary focus-border-primary focus:outline-none transition-colors ${
-                        errors.channels && errors.channels[i] ? 'border-red-500' : ''
-                      }`}
-                    />
-                    {errors.channels && errors.channels[i] && (
-                      <div className="mt-1 flex items-center text-red-500 text-sm">
-                        <AlertCircle size={14} className="mr-1" />
-                        <span>{errors.channels[i]}</span>
-                      </div>
-                    )}
-                  </>
-                )}
-              </div>
-
-              {channels.length > 1 && (
-                <Button variant="ghost" size="icon" onClick={() => removeChannel(i)} className="text-red-500 h-8 w-8">
-                  <X size={16} />
-                </Button>
-              )}
-            </div>
-          ))}
-        </CardContent>
-      </Card>
+      <DeliveryChannelsCard
+        channels={channels}
+        addChannel={addChannel}
+        updateChannel={updateChannel}
+        removeChannel={removeChannel}
+        errors={errors}
+        session={session}
+      />
 
       {/* Action Buttons */}
       <div className="flex justify-between">
